Execute consumers in separate thread
Fix a bug that prevents Pika from sending heartbeat messages to the message broker. When a long-running consumer blocks the main thread, Pika wouldn't send any heartbeat messages during this time and the RabbitMQ server will kill the connection as a consequence.

Handler execution has been moved to a separate thread pool. Because Pika's connection and channel objects are not threadsafe, use the `add_callback_threadsafe` that has been introduced with Pika 0.12. If a handler throws an exception, the consuming process is not killed anymore.

Python 3 has the sanest thread pool implementation, that's why I used that and dropped Python 2 support for future version of this library.