Avoid connection errors during interactive replay
When using replay interactively, it's likely that the connection is lost
while one is investigating why the message may have been dead-lettered
and whether it should be replayed or discarded.

Use a timeout while waiting for user input. The timeout must be smaller
than the timouts or intervals set at the RabbitMQ broker.